+++ CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- The `IRequestCacheControl` interface was renamed to `ICacheControl`.
+
+### Added
+
+- The `AllowCacheUse` and `AllowCacheUpdate` extension methods were added
+  to simplify the otherwise required negations of `NoCache` and `NoStore`.
+
 ## 0.3.0 - 2022-11-09
 
 ### Added